Remove authorizer and directory subcommands
