--- pkg/cli/cmd/authorizer/decisiontree.go
@@ -1,47 +1,0 @@
-package authorizer
-
-import (
-	"github.com/aserto-dev/topaz/pkg/cli/cc"
-	"github.com/aserto-dev/topaz/pkg/cli/clients"
-	"github.com/aserto-dev/topaz/pkg/cli/jsonx"
-	"github.com/pkg/errors"
-
-	"github.com/aserto-dev/go-authorizer/aserto/authorizer/v2"
-	"github.com/aserto-dev/go-authorizer/aserto/authorizer/v2/api"
-)
-
-type DecisionTreeCmd struct {
-	AuthParams     `embed:""`
-	Path           string   `name:"path" help:"policy package to evaluate"`
-	Decisions      []string `name:"decisions" default:"*" help:"policy decisions to return"`
-	clients.Config `envprefix:"TOPAZ_AUTHORIZER_"`
-}
-
-func (cmd *DecisionTreeCmd) Run(c *cc.CommonCtx) error {
-	client, err := clients.NewAuthorizerClient(c, &cmd.Config)
-	if err != nil {
-		return errors.Wrap(err, "failed to get authorizer client")
-	}
-
-	resource, err := cmd.ResourceContext()
-	if err != nil {
-		return err
-	}
-
-	resp, err := client.DecisionTree(c.Context, &authorizer.DecisionTreeRequest{
-		PolicyContext: &api.PolicyContext{
-			Path:      cmd.Path,
-			Decisions: cmd.Decisions,
-		},
-		IdentityContext: cmd.IdentityContext(),
-		ResourceContext: resource,
-		Options: &authorizer.DecisionTreeOptions{
-			PathSeparator: authorizer.PathSeparator_PATH_SEPARATOR_DOT,
-		},
-	})
-	if err != nil {
-		return err
-	}
-
-	return jsonx.OutputJSONPB(c.UI.Output(), resp)
-}
