+++ pkg/cli/cmd/exporter.go
@@ -1,4 +1,4 @@
-package directory
+package cmd
 
 import (
 	"path/filepath"
@@ -10,9 +10,9 @@ import (
 )
 
 type ExportCmd struct {
-	Directory      string        `short:"d" required:"" help:"directory to write .json data"`
-	Format         FormatVersion `flag:"" short:"f" enum:"3,2" name:"format" default:"3" help:"format of json data"`
-	clients.Config `envprefix:"TOPAZ_DIRECTORY_"`
+	Directory string        `short:"d" required:"" help:"directory to write .json data"`
+	Format    FormatVersion `flag:"" short:"f" enum:"3,2" name:"format" default:"3" help:"format of json data"`
+	clients.Config
 }
 
 func (cmd *ExportCmd) Run(c *cc.CommonCtx) error {
