+++ pkg/cli/dockerx/docker.go
@@ -11,12 +11,14 @@ import (
 	"strings"
 	"syscall"
 
+	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/filters"
 	"github.com/docker/docker/api/types/image"
 	"github.com/docker/docker/api/types/network"
 	"github.com/docker/docker/client"
 	"github.com/docker/go-connections/nat"
+
 	v1 "github.com/opencontainers/image-spec/specs-go/v1"
 )
